Refactor backend: move classes into api, domain, repository packages
