+++ backend/src/main/java/fi/fisma/backend/repository/AppUserRepository.java
@@ -1,6 +1,8 @@
-package fi.fisma.backend.appuser;
+package fi.fisma.backend.repository;
 
 import org.springframework.data.repository.ListCrudRepository;
+
+import fi.fisma.backend.domain.AppUser;
 
 public interface AppUserRepository extends ListCrudRepository<AppUser, Long> {
     AppUser findByUsername(String username);
