+++ backend/src/test/java/fi/fisma/backend/domain/AppUserJsonTest.java
@@ -1,4 +1,4 @@
-package fi.fisma.backend.appuser;
+package fi.fisma.backend.domain;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
