+++ backend/src/test/java/fi/fisma/backend/domain/ProjectJsonTest.java
@@ -1,11 +1,10 @@
-package fi.fisma.backend.project;
+package fi.fisma.backend.domain;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.json.JsonTest;
 import org.springframework.boot.test.json.JacksonTester;
-
 import java.time.LocalDateTime;
 import java.util.Set;
